Move things out of extras package
- Move ExtractInstant to core.data
- Move CaptureTimeRange to test code example package
