+++ core/src/test/scala/com/rallyhealth/vapors/core/example/CaptureTimeRange.scala
@@ -1,8 +1,29 @@
-package com.rallyhealth.vapors.factfilter.extras
+package com.rallyhealth.vapors.core.example
+
+import cats.{Eval, Monoid}
+import com.rallyhealth.vapors.core.algebra.{CaptureP, Expr}
+import com.rallyhealth.vapors.core.data.{ExtractInstant, TypedFact}
+import com.rallyhealth.vapors.core.interpreter.{ExprInput, ExprOutput}
 
 import java.time.Instant
 
-import cats.Monoid
+object CaptureTimeRange extends CaptureP.AsMonoidCompanion[TimeRange] {
+
+  implicit def captureTimeRangeFromFacts[T : ExtractInstant, R]: CaptureP.AsMonoidFromFactsOfType[T, R, TimeRange] = {
+    new CaptureP.AsMonoidFromFactsOfType[T, R, TimeRange] {
+
+      override protected def foldWithParentParam(
+        expr: Expr[Seq, TypedFact[T], R, TimeRange],
+        input: ExprInput[Seq, TypedFact[T]],
+        output: ExprOutput[R],
+        processedChildren: TimeRange,
+      ): Eval[TimeRange] = {
+        val timestamps = input.value.map(fact => ExtractInstant[T].extractValue(fact.value))
+        Eval.now(TimeRange.fromIterable(timestamps))
+      }
+    }
+  }
+}
 
 /**
   * A range of time that tracks the start and end of the range, without any concern for the values inbetween.
